pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "openai>=1.92.2",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]
